ron - fix relationships bug (not possible to delete connected objects) using fk constraint also adding forms validations.

+++ GRD/GRD/Controllers/BranchesController.cs
@@ -160,6 +160,11 @@ namespace GRD.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var branch = await _context.Branches.FindAsync(id);
+            if (branch == null)
+            {
+                return View("Views/Branches/NotFound.cshtml");
+            }
+            branch.Purchases.ToList().ForEach(p => branch.Purchases.Remove(p));
             _context.Branches.Remove(branch);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
